requirements.txt
flask
flask_socketio
openai
google-cloud-speech
scipy
pyaudio
librosa
numpy
pylint
eslint